requirements.txt
numpy==1.26.2
pandas==2.1.3
python-binance==1.0.19
tensorflow==2.15.0
scikit-learn==1.4.1.post1
grpcio==1.65.1
grpcio-tools
protobuf
pymongo
redis
python-dotenv